migration/migration_engine.py
import asyncio
import logging
import os
import pymssql
from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
from sqlalchemy.orm import sessionmaker
from dotenv import load_dotenv

load_dotenv()

logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

# Legacy MSSQL Credentials (from .env)
MSSQL_HOST = os.getenv("MSSQL_HOST", "localhost")
MSSQL_USER = os.getenv("MSSQL_USER", "sa")
MSSQL_PASS = os.getenv("MSSQL_PASS", "password")
MSSQL_DB = os.getenv("MSSQL_DB", "legacy_db")

# VetOS PostgreSQL Credentials
POSTGRES_URL = os.getenv("VETOS_POSTGRES_URL", "postgresql+asyncpg://vetos_user:vetos_password@localhost:5432/vetos_db")

class VetOSMigrationEngine:
    def __init__(self):
        self.pg_engine = create_async_engine(POSTGRES_URL, echo=False)
        self.pg_session = sessionmaker(self.pg_engine, class_=AsyncSession, expire_on_commit=False)
        self.mssql_conn = None

    def connect_mssql(self):
        try:
            self.mssql_conn = pymssql.connect(MSSQL_HOST, MSSQL_USER, MSSQL_PASS, MSSQL_DB)
            logger.info("Connected to Legacy MSSQL.")
            return True
        except Exception as e:
            logger.error(f"Failed to connect to MSSQL: {e}")
            return False

    async def migrate_clients(self):
        logger.info("Extracting clients from MSSQL...")
        # TODO: Implement mapping after RO access
        # cursor = self.mssql_conn.cursor(as_dict=True)
        # cursor.execute("SELECT ... FROM ...")
        pass

    async def migrate_patients(self):
        logger.info("Extracting patients from MSSQL...")
        # TODO: Implement mapping after RO access
        pass

    async def run(self):
        if not self.connect_mssql():
            return
        
        try:
            await self.migrate_clients()
            await self.migrate_patients()
            logger.info("Migration cycle completed successfully.")
        finally:
            if self.mssql_conn:
                self.mssql_conn.close()

async def main():
    engine = VetOSMigrationEngine()
    await engine.run()

if __name__ == "__main__":
    asyncio.run(main())
