migration/requirements.txt
pymssql
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
python-dotenv
